Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/cssom/cssom-fontfacerule-expected.txt


FAIL CSSStyleDeclaration values are represented within CSSFontFaceRule assert_equals: expected "url(\"http://foo/bar/font.ttf\")" but got "url(http://foo/bar/font.ttf)"
